+++ bom/pom.xml
@@ -93,6 +93,11 @@
         <artifactId>zedai-to-ssml</artifactId>
         <version>1.0.0</version>
       </dependency>
+      <dependency>
+        <groupId>org.daisy.pipeline.modules</groupId>
+        <artifactId>tts-helpers</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+      </dependency>
     </dependencies>
   </dependencyManagement>
 </project>
