+++ css-speech/pom.xml
@@ -40,7 +40,7 @@
     <!-- needed for the ConfigReader -->
     <dependency>
       <groupId>org.daisy.pipeline.modules</groupId>
-      <artifactId>tts-common</artifactId>
+      <artifactId>tts-helpers</artifactId>
     </dependency>
     <dependency>
      <groupId>junit</groupId>
