+++ ssml-to-audio/pom.xml
@@ -22,6 +22,10 @@
       <groupId>org.daisy.pipeline.modules</groupId>
       <artifactId>tts-common</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.daisy.pipeline.modules</groupId>
+      <artifactId>tts-helpers</artifactId>
+    </dependency>
     <!-- needed to use com.xmlcalabash.library.DefaultStep ... -->
     <dependency>
       <groupId>org.daisy.libs</groupId>
