+++ ssml-to-audio/src/main/resources/META-INF/catalog.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <catalog xmlns="urn:oasis:names:tc:entity:xmlns:xml:catalog">
-  
+
   <uri name="http://www.daisy.org/pipeline/modules/ssml-to-audio/library.xpl"
        uri="../xml/library.xpl"/>
 
   <nextCatalog catalog="org:daisy:pipeline:modules:common-utils"/>
   <nextCatalog catalog="org:daisy:pipeline:modules:fileset-utils" />
+  <nextCatalog catalog="org:daisy:pipeline:modules:tts-helpers" />
 
 </catalog>
