+++ tts-utils/tts-helpers/src/main/resources/META-INF/catalog.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<catalog xmlns="urn:oasis:names:tc:entity:xmlns:xml:catalog">
+
+  <uri name="http://www.daisy.org/pipeline/modules/tts-helpers/library.xpl"
+       uri="../xml/xproc/library.xpl"/>
+
+  <nextCatalog catalog="org:daisy:pipeline:modules:fileset-utils"/>
+</catalog>
