+++ tts-utils/tts-helpers/src/main/resources/xml/xproc/library.xpl
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <p:library xmlns:p="http://www.w3.org/ns/xproc" version="1.0">
 
-    <p:import href="ssml-to-audio.xpl"/>
+    <p:import href="create-audio-fileset.xpl"/>
 
 </p:library>
